Removed select_utxos_textkeyed.
Removed some debug statements.
Fixed a detail in the maker side communication between client
and daemon: because json only allows strings for keys, and does
not allow binary, created a translation within the processing
in maker client protocol to ensure that the utxos dict could be
correctly passed across.
Fixed a couple of minor bugs.